requirements.txt
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-admin-select2==1.0.1
django-background-tasks==1.2.5
django-captcha-admin==0.2
django-compat==1.0.15
django-htmlmin==0.11.0
django-lazysignup==2.0.0
django-recaptcha==1.3.0
facebook-sdk==3.1.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
html5lib==1.0.1
httplib2==0.17.0
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.7.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
pytz==2019.3
razorpay==1.2.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
